requirements.txt
asyncio-nats-client==0.3.0